release v0.1 => optimized for full functioning after initial commit

+++ pseudo_proccessor_main.sh
@@ -1,5 +1,4 @@
-#!/bin/bash/
-
+#!/bin/bash
 # conda activate "bacterial-genomics-tutorial"
 
 cpus=$2
@@ -7,11 +6,12 @@ cpus=$2
 # echo "Using $cpus threads"
 base_dir=$1
 seqid=${1##*/}
-organism_tag=$5
-id="${organism_tag}${seqid:3}"
+id="$5"
 dnds=$6
 scaffold=$7
 predictor_type=$8
+stage5=$9
+app_dir=${0%/*}
 
 if [[ "$4" != "''" ]]; then ancestor_genome_gbff="$4.gbff"; ancestor_genome_fna="$4.fna"; else ancestor_genome_gbff=""; ancestor_genome_fna=""; fi
 if [[ "$3" != "''"  ]]; then reference_genome_gbff="$3.gbff"; reference_genome_fna="$3.fna"; else reference_genome_gbff=""; reference_genome_fna=""; fi
@@ -22,17 +22,17 @@ ancestor_genome_gbff=$ancestor_genome_gbff
 ancestor_genome_fna=$ancestor_genome_fna
 base_dir=$1
 seqid=${1##*/}
-organism_tag=$5
-id="${organism_tag}${seqid:3}"
+id=$5
 dnds=$6
 scaffold=$7
-predictor_type=$8\n\n"
+predictor_type=$8
+stage5=$stage5\n\n"
 
 prokka_pseudos_annotator=./prokka_anno.pl
 
 mkdir -p "$base_dir/rt_output"
 
-if [[ "$scaffold" = "T" ]]; then 
+if [[ "$scaffold" = "T" ]] || [[ "$stage5" = "a" ]]; then 
     mkdir -p "$base_dir/rt_output/corrected"
 
     echo -e "Correcting misassemblies with ragtag\n" | tee -a "$base_dir/pseudogenome_processing.txt"
@@ -52,15 +52,19 @@ if [[ "$scaffold" = "T" ]]; then
     # cp "$base_dir/rt_output/ragtag.scaffold.fasta" "$base_dir/rt_output/corrected/ragtag.scaffold.fasta" 
     #KP work problems with ragtag solution
     sed -E '/^>/!s/[^ATCGN]*(>|[^ATCGN]).?(>|[^ATCGN])//g' "$base_dir/rt_output/ragtag.scaffold.fasta" > "$base_dir/rt_output/ragtag.scaffold_corrected.fasta"
+    #sed -E '/^>/!s/[^ATCGN]*(>|[^ATCGN]).?(>|[^ATCGN])//g' "$base_dir/rt_output/ragtag.scaffold_corrected.fasta" > "$base_dir/rt_output/ragtag.scaffold_corrected.fasta"
+    # cp "$base_dir/rt_output/ragtag.scaffold.fasta" "$base_dir/rt_output/ragtag.scaffold_corrected.fasta"
     echo -e "################################################################################################\n"
 elif [[ "$scaffold" = "F" ]]; then
-    cp -f "$base_dir/${seqid}.fna" "$base_dir/rt_output/ragtag.scaffold_corrected.fasta"
+    if [ -f "$base_dir/${seqid}.fna" ]; then
+        cp -f "$base_dir/${seqid}.fna" "$base_dir/rt_output/ragtag.scaffold_corrected.fasta"
+    fi
 fi
 
 # # echo -e "Genome assembly QC on final assembly\n"
 # # start_assembly_qc=`date +%s`
-# # # ./assembly_qc_scale.sh $base_dir $cpus $assembly_qc_method
-# # ./assembly_qc_scale_U50_QUAST.sh $base_dir $cpus $assembly_qc_method $organism_tag
+# # # ./assembly_qc.sh $base_dir $cpus $assembly_qc_method
+# # ./assembly_qc_scale_U50_QUAST.sh $base_dir $cpus $assembly_qc_method $id
 # # end_assembly_qc=`date +%s`
 # # runtime=$((end_assembly_qc-start_assembly_qc))
 # # hours=$((runtime / 3600)); minutes=$(( (runtime % 3600) / 60 )); seconds=$(( (runtime % 3600) % 60 )) 
@@ -95,6 +99,7 @@ function prokka_pseudo_annotation () {
 
 #Annotating with DFAST
 function DFAST_annotation () {
+    conda activate "bactopia_manually"
     echo -e "Annotating with DFAST\n" | tee -a "$base_dir/pseudogenome_processing.txt"
     # conda activate 'DFAST'
     if [[ $ancestor_genome_gbff ]]; then
@@ -115,12 +120,12 @@ function DFAST_annotation () {
     echo -e "################################################################################################\n"
 }
 
-Pseudogene prediction with pseudofinder
+#Pseudogene prediction with pseudofinder
 function pseudofinder_prediction () {
     echo -e "Predicting pseudogenes with Pseudofinder using Prokka annotation\n" | tee -a "$base_dir/pseudogenome_processing.txt"
     conda activate "pseudofinder"
     start_PF=`date +%s`
-    if [[ $ancestor_genome_gbff ]]; then
+    if [[ $ancestor_genome_gbff ]] && [[ $dnds ]]; then
         ~/pseudofinder256/pseudofinder.py annotate --genome "${seqid}/rt_output/prokka_annotation/${id}.gbk" --outprefix "${id}" \
             --database "~/pseudofinder256/swissprot.dmnd" --threads $cpus --diamond \
             -ref $ancestor_genome_gbff \
@@ -143,48 +148,73 @@ function pseudofinder_prediction () {
     echo -e "################################################################################################\n"
 }
 
-if [[ "$predictor_type" = "prokka" ]]; then
+#Processing of predicted pseudogenes
+function pseudogenes_extractor () {
+    echo -e "Extracting predicted pseudogenes from $1\n" | tee -a "$base_dir/pseudogenome_processing.txt"
+    # conda activate "pseudofinder"
+    pseudogene_processor="${app_dir}/pseudo_proccessor_sub.py"
+    DFAST_gbk="$base_dir/rt_output/DFAST_output/genome.gbk"
+    prokka_gbk="$base_dir/rt_output/prokka_annotation/$id.gbk"
+    PF_gff="$base_dir/pseudofinder/PF_output_pseudos.gff"
+    prokka_pseudos_tsv="$base_dir/rt_output/prokka_annotation/$id.faa.prokka_pseudos.tsv"
+    outdir="$base_dir/pseudogenes"
+    mkdir -p $outdir
+
+    python $pseudogene_processor \
+        $DFAST_gbk \
+        $prokka_gbk \
+        $PF_gff \
+        $prokka_pseudos_tsv \
+        $outdir $predictor_type\
+        1> $base_dir/pseudogene_processor.stdout.txt 2> $base_dir/pseudogene_processor.stderr.txt
+    [ -s $base_dir/pseudogene_processor.stderr.txt ] && cat $base_dir/pseudogene_processor.stderr.txt || \
+        echo -e "Predicted pseudogenes extracted successfully\n" | tee -a "$base_dir/pseudogenome_processing.txt"
+    echo -e "################################################################################################\n"
+}
+
+#Removing duplicates with CDHIT
+function cleaner_cdhit () {
+    conda activate "bactopia_manually"
+    echo -e "Removing duplicates with CDHIT\n" | tee -a "$base_dir/pseudogenome_processing.txt"
+    mkdir -p "$base_dir/cdhit" 
+    cd-hit -i "$base_dir/pseudogenes/combined_pseudos.fasta" -o "$base_dir/cdhit/${id}_cdhit" -c 0.95 -n 5 -g 1 -M 6000 -T $cpus -A 0.95 -s 0.9 \
+        1> $base_dir/cd_hit.stdout.txt 2> $base_dir/cd_hit.stderr.txt
+    [ -s $base_dir/cd_hit.stderr.txt ] && cat $base_dir/cd_hit.stderr.txt || \
+        cp "$base_dir/cdhit/${id}_cdhit" "$base_dir/cdhit/${id}_cdhit.fasta"
+    grep ">" "$base_dir/cdhit/${id}_cdhit.fasta" | wc -l | echo -e "$(</dev/stdin) unique pseudogenes found\n" | tee -a "$base_dir/pseudogenome_processing.txt"
+}
+
+if [[ "$stage5" = "b" ]]; then
+    prokka_annotation
+elif [[ "$stage5" = "c" ]]; then
+    prokka_pseudo_annotation
+elif [[ "$stage5" = "d" ]]; then
+    DFAST_annotation
+elif [[ "$stage5" = "e" ]]; then
+    pseudofinder_prediction
+elif [[ "$stage5" = "f" ]]; then
+    pseudogenes_extractor
+elif [[ "$stage5" = "g" ]]; then
+    cleaner_cdhit
+elif [[ "$predictor_type" = "prokka" ]]; then
     prokka_annotation
     prokka_pseudo_annotation
+    pseudogenes_extractor $predictor_type
+    cleaner_cdhit
 elif [[ "$predictor_type" = "PF" ]]; then
     prokka_annotation
     pseudofinder_prediction
+    pseudogenes_extractor $predictor_type
+    cleaner_cdhit
 elif [[ "$predictor_type" = "DFAST" ]]; then
     DFAST_annotation
+    pseudogenes_extractor $predictor_type
+    cleaner_cdhit
 elif [[ "$predictor_type" = "all" ]]; then
     prokka_annotation
     prokka_pseudo_annotation
     pseudofinder_prediction
     DFAST_annotation
+    pseudogenes_extractor $predictor_type
+    cleaner_cdhit
 fi
-
-#Processing of predicted pseudogenes
-echo -e "Extracting predicted pseudogenes from all tools\n" | tee -a "$base_dir/pseudogenome_processing.txt"
-# conda activate "pseudofinder"
-pseudogene_processor="pseudo_proccessor_sub.py"
-DFAST_gbk="$base_dir/rt_output/DFAST_output/genome.gbk"
-prokka_gbk="$base_dir/rt_output/prokka_annotation/$id.gbk"
-PF_gff="$base_dir/pseudofinder/PF_output_pseudos.gff"
-prokka_pseudos_tsv="$base_dir/rt_output/prokka_annotation/$id.faa.prokka_pseudos.tsv"
-outdir="$base_dir/pseudogenes"
-mkdir -p $outdir
-python $pseudogene_processor \
-    $DFAST_gbk \
-    $prokka_gbk \
-    $PF_gff \
-    $prokka_pseudos_tsv \
-    $outdir \
-    1> $base_dir/pseudogene_processor.stdout.txt 2> $base_dir/pseudogene_processor.stderr.txt
-[ -s $base_dir/pseudogene_processor.stderr.txt ] && cat $base_dir/pseudogene_processor.stderr.txt || \
-    echo -e "Predicted pseudogenes extracted successfully\n" | tee -a "$base_dir/pseudogenome_processing.txt"
-echo -e "################################################################################################\n"
-
-#Removing duplicates with CDHIT
-conda activate "bactopia_manually"
-echo -e "Removing duplicates with CDHIT\n" | tee -a "$base_dir/pseudogenome_processing.txt"
-mkdir -p "$base_dir/cdhit" 
-cd-hit -i "$outdir/combined_pseudos.fasta" -o "$base_dir/cdhit/${id}_cdhit" -c 0.95 -n 5 -g 1 -M 6000 -T $cpus -A 0.95 -s 0.9 \
-    1> $base_dir/cd_hit.stdout.txt 2> $base_dir/cd_hit.stderr.txt
-[ -s $base_dir/cd_hit.stderr.txt ] && cat $base_dir/cd_hit.stderr.txt || \
-    cp "$base_dir/cdhit/${id}_cdhit" "$base_dir/cdhit/${id}_cdhit.fasta"
-grep ">" "$base_dir/cdhit/${id}_cdhit.fasta" | wc -l | echo -e "$(</dev/stdin) unique pseudogenes found\n" | tee -a "$base_dir/pseudogenome_processing.txt"
